translator/requirements.txt
python-telegram-bot==24.6
requests==2.26.0
